requirements.txt
bcrypt==4.1.3
click==8.1.7
mypy==1.10.0
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
prettytable==3.10.0
python-dotenv==1.0.1
typing_extensions==4.12.2
wcwidth==0.2.13
